# Try masked fill (but stick with ol' multiply and add comment)