Tried to find timings to support making the change but it turns out it's mostly identical or slower. 

When called on square matrix it's near identical (marginally slower maybe 2%)
<details> <summary>logs</summary>

```
# CUDA T4 (512 x 512)
# master
# args.no_cuda: False
# GPU: True
# torch.__version__: 2.5.1+cu121
# eval_mode:     False     requires_grad: True      backward: True      
# 0m 25s    (   25.247 s total )     TailDropout
# 0m 24s    (   24.740 s total )     TailDropout
# eval_mode:     False     requires_grad: True      backward: False     
# 0m 0s     (    0.989 s total )     TailDropout
# 0m 1s     (    1.176 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: True      
# 0m 22s    (   22.785 s total )     TailDropout
# 0m 22s    (   22.843 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: False     
# 0m 0s     (    0.221 s total )     TailDropout
# 0m 0s     (    0.214 s total )     TailDropout
# FINISH, total time: 1m 38s (98s tot)

# This
# args.no_cuda: False
# GPU: True
# torch.__version__: 2.5.1+cu121
# eval_mode:     False     requires_grad: True      backward: True      
# 0m 24s    (   24.848 s total )     TailDropout
# 0m 24s    (   24.922 s total )     TailDropout
# eval_mode:     False     requires_grad: True      backward: False     
# 0m 0s     (    0.998 s total )     TailDropout
# 0m 0s     (    0.993 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: True      
# 0m 22s    (   22.926 s total )     TailDropout
# 0m 23s    (   23.020 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: False     
# 0m 0s     (    0.205 s total )     TailDropout
# 0m 0s     (    0.226 s total )     TailDropout
# FINISH, total time: 1m 38s (98s tot)

# CPU (T4) (128 x 128)
# master
# args.no_cuda: True
# GPU: False
# torch.__version__: 2.5.1+cu121
# eval_mode:     False     requires_grad: True      backward: True      
# 0m 17s    (   17.601 s total )     TailDropout
# 0m 17s    (   17.784 s total )     TailDropout
# eval_mode:     False     requires_grad: True      backward: False     
# 0m 2s     (    2.836 s total )     TailDropout
# 0m 2s     (    2.173 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: True      
# 0m 11s    (   11.239 s total )     TailDropout
# 0m 11s    (   11.635 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: False     
# 0m 0s     (    0.367 s total )     TailDropout
# 0m 0s     (    0.240 s total )     TailDropout
# FINISH, total time: 1m 3s (64s tot)

# This
# args.no_cuda: True
# GPU: False
# torch.__version__: 2.5.1+cu121
# eval_mode:     False     requires_grad: True      backward: True      
# 0m 17s    (   17.569 s total )     TailDropout
# 0m 18s    (   18.646 s total )     TailDropout
# eval_mode:     False     requires_grad: True      backward: False     
# 0m 2s     (    2.463 s total )     TailDropout
# 0m 2s     (    2.135 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: True      
# 0m 11s    (   11.293 s total )     TailDropout
# 0m 11s    (   11.415 s total )     TailDropout
# eval_mode:     True      requires_grad: True      backward: False     
# 0m 0s     (    0.210 s total )     TailDropout
# 0m 0s     (    0.215 s total )     TailDropout
# FINISH, total time: 1m 3s (64s tot)

# My Machine (MBP)
# master
# args.no_cuda: False
# GPU: False
# torch.__version__: 2.2.2
# eval_mode: False 	 equires_grad:  True 	 backward:  True
# 0m 20s 	( 20.728883266448975 	 s total)	 TailDropout
# 0m 22s 	( 22.256752252578735 	 s total)	 TailDropout
# eval_mode: False 	 equires_grad:  True 	 backward:  False
# 0m 13s 	( 13.129317045211792 	 s total)	 TailDropout
# 0m 11s 	( 11.208441734313965 	 s total)	 TailDropout
# eval_mode: True 	 equires_grad:  True 	 backward:  True
# 0m 8s 	( 8.223066806793213 	 s total)	 TailDropout
# 0m 8s 	( 8.199806928634644 	 s total)	 TailDropout
# eval_mode: True 	 equires_grad:  True 	 backward:  False
# 0m 0s 	( 0.0019350051879882812 	 s total)	 TailDropout
# 0m 0s 	( 0.0017910003662109375 	 s total)	 TailDropout
# FINISH, total time: 3m 17s (198s tot)

# This
# torch.__version__: 2.2.2
# eval_mode: False 	 equires_grad:  True 	 backward:  True
# 0m 25s 	( 25.14108896255493 	 s total)	 TailDropout
# 0m 21s 	( 21.887511253356934 	 s total)	 TailDropout
# eval_mode: False 	 equires_grad:  True 	 backward:  False
# 0m 8s 	( 8.231560945510864 	 s total)	 TailDropout
# 0m 7s 	( 7.942879915237427 	 s total)	 TailDropout
# eval_mode: True 	 equires_grad:  True 	 backward:  True
# 0m 8s 	( 8.660022020339966 	 s total)	 TailDropout
# 0m 9s 	( 9.132021188735962 	 s total)	 TailDropout
# eval_mode: True 	 equires_grad:  True 	 backward:  False
# 0m 0s 	( 0.0016608238220214844 	 s total)	 TailDropout
# 0m 0s 	( 0.0016109943389892578 	 s total)	 TailDropout
# FINISH, total time: 3m 22s (203s tot)
```
</details>

But when applied to non-square shapes like `[512, 10, 10, 512]` it starts to be slower.

Both are almost identical when compiled
```
# 3.85 s ± 3.76 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # master compiled
# 3.33 s ± 26.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # master not compiled

# 5.21 s ± 181 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # this   not compiled
# 4.96 s ± 20.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # this   not compiled
# 4.99 s ± 42.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
# 3.85 s ± 4.11 ms per loop (mean ± std. dev. of 7 runs, 1 loop each) # this   compiled
```
<details>

```
import torch
import torch.nn as nn
import torch.nn.functional as F

import time
import math
class Args:  # A dummy class
    repeats = 1000
    n_features = 512
    batch_size = 512
    no_cuda = False
    device = 'cuda'
    time_limit = None
    requires_grad =True
    backward = True

args = Args()
dropout = TailDropout().to('cuda')

# dropout = torch.compile(dropout)

y = torch.ones([args.batch_size, 10, 10, args.n_features],requires_grad=args.requires_grad, device=args.device)
dropout(y).shape

%%timeit
optimizer = torch.optim.SGD((y,), lr=0.1)

# Work
start = time.time()
for _ in range(args.repeats):
    z = dropout(y)
    loss = z.sum()
    loss.backward()
    optimizer.step()
    optimizer.zero_grad()

```
<details>


This PR documents this in comments and tiny fixs